+++ includes/task.h
@@ -2,10 +2,11 @@
 #define TASK_H
 
 #include <stdint.h>
+#include <stdlib.h>
 
 #define TASK_LIMIT 16
 
-void    task_add(void (*task_code)(void), uint16_t stack_size);
+void    task_add(void (*task_code)(void), size_t stack_size);
 void    task_scheduler(void);
 
 #endif
